Refactored app
